Feat: 차들을 받아 경주를 시키고 우승자를 구하는 레이싱 시스템으로서의 Racing 클래스 생성 및 구현
